+++ mybatis-plus-dts/src/main/java/com/baomidou/mybatisplus/dts/EnableRmtRabbit.java
@@ -1,6 +1,7 @@
-package com.baomidou.mybatisplus.rmt.config;
+package com.baomidou.mybatisplus.dts;
 
-import com.baomidou.mybatisplus.rmt.mq.RabbitConfiguration;
+import com.baomidou.mybatisplus.dts.rabbit.config.RmtAutoConfiguration;
+import com.baomidou.mybatisplus.dts.rabbit.mq.RabbitConfiguration;
 import org.springframework.context.annotation.Import;
 
 import java.lang.annotation.Documented;
@@ -19,7 +20,7 @@ import java.lang.annotation.Target;
 @Retention(value = java.lang.annotation.RetentionPolicy.RUNTIME)
 @Target(value = { java.lang.annotation.ElementType.TYPE })
 @Documented
-@Import({RabbitConfiguration.class})
+@Import({RabbitConfiguration.class, RmtAutoConfiguration.class})
 public @interface EnableRmtRabbit {
 
 }
