+++ mybatis-plus-dts/src/main/java/com/baomidou/mybatisplus/dts/rabbit/RmtMeta.java
@@ -1,4 +1,4 @@
-package com.baomidou.mybatisplus.rmt;
+package com.baomidou.mybatisplus.dts.rabbit;
 
 import lombok.Data;
