--- mybatis-plus-rmt/src/main/java/com/baomidou/mybatisplus/rmt/daemon/RmtResendProcess.java
@@ -1,11 +1,0 @@
-package com.baomidou.mybatisplus.rmt.daemon;
-
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class RmtResendProcess {
-
-    // 处理任务重发
-
-}
