+++ settings.gradle
@@ -1,8 +1,8 @@
 rootProject.name = 'mybatis-plus-root'
 include 'mybatis-plus'
 include 'mybatis-plus-core'
+include 'mybatis-plus-dts'
 include 'mybatis-plus-annotation'
 include 'mybatis-plus-extension'
 include 'mybatis-plus-generator'
-include 'mybatis-plus-rmt'
 include 'mybatis-plus-boot-starter'
